gcp/predict/requirements.txt
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
numpy
mlflow
db-dtypes
keras